requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.0.0=pypi_0
argon2-cffi=21.1.0=py39hb82d6ee_2
astunparse=1.6.3=pypi_0
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
blas=1.0=mkl
bleach=4.1.0=pyhd8ed1ab_0
brotli=1.0.9=ha925a31_2
ca-certificates=2021.10.26=haa95532_2
cachetools=4.2.4=pypi_0
certifi=2021.10.8=py39haa95532_0
cffi=1.15.0=py39h0878f49_0
charset-normalizer=2.0.9=pypi_0
colorama=0.4.4=pyh9f0ad1d_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.5.1=py39h415ef7b_0
decorator=5.1.0=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
flatbuffers=2.0=pypi_0
fonttools=4.28.3=pypi_0
freetype=2.10.4=hd328e21_0
gast=0.4.0=pypi_0
google-auth=2.3.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.42.0=pypi_0
h5py=3.6.0=pypi_0
icu=58.2=ha925a31_3
idna=3.3=pypi_0
importlib-metadata=4.8.2=py39hcbf5309_0
importlib_resources=5.4.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.6.0=py39h832f523_0
ipython=7.30.1=py39hcbf5309_0
ipython_genutils=0.2.0=py_1
jedi=0.18.1=py39hcbf5309_0
jinja2=3.0.3=pyhd8ed1ab_0
joblib=1.1.0=pypi_0
jpeg=9d=h2bbff1b_0
jsonschema=4.2.1=pyhd8ed1ab_1
jupyter_client=7.1.0=pyhd8ed1ab_0
jupyter_core=4.9.1=py39hcbf5309_1
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
keras=2.7.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.2=pypi_0
libclang=12.0.0=pypi_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.18=h8d14728_1
libtiff=4.2.0=hd0e1b90_0
libwebp=1.2.0=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.6=pypi_0
markupsafe=2.0.1=py39hb82d6ee_1
matplotlib=3.5.0=pypi_0
matplotlib-base=3.5.0=py39h6214cd6_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mediapipe=0.8.9=pypi_0
mistune=0.8.4=py39hb82d6ee_1005
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
msys2-conda-epoch=20160418=1
munkres=1.1.4=py_0
nbclient=0.5.9=pyhd8ed1ab_0
nbconvert=6.3.0=py39hcbf5309_1
nbformat=5.1.3=pyhd8ed1ab_0
nest-asyncio=1.5.4=pyhd8ed1ab_0
notebook=6.4.6=pyha770c72_0
numpy=1.21.4=pypi_0
numpy-base=1.21.2=py39h0829f74_0
oauthlib=3.1.1=pypi_0
olefile=0.46=pyhd3eb1b0_0
opencv-contrib-python=4.5.4.60=pypi_0
opencv-python=4.5.4.60=pypi_0
openssl=1.1.1l=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=pypi_0
pandoc=2.16.2=h8ffe710_0
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=8.4.0=pypi_0
pip=21.2.4=py39haa95532_0
prometheus_client=0.12.0=pyhd8ed1ab_0
prompt-toolkit=3.0.24=pyha770c72_0
protobuf=3.19.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.10.0=pyhd8ed1ab_0
pyparsing=3.0.6=pyhd8ed1ab_0
pyqt=5.9.2=py39hd77b12b_6
pyrsistent=0.18.0=py39hb82d6ee_0
python=3.9.7=h6244533_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.3=pypi_0
pywin32=302=py39hb82d6ee_2
pywinpty=0.5.7=py39haa95532_0
pyzmq=22.3.0=py39he46f08e_1
qt=5.9.7=vc14h73c81de_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.1=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=58.0.4=py39haa95532_0
setuptools-scm=6.3.2=pypi_0
sip=4.19.13=py39hd77b12b_0
six=1.16.0=pyh6c4a22f_0
spotipy=2.19.0=pypi_0
sqlite=3.36.0=h2bbff1b_0
tensorboard=2.7.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.7.0=pypi_0
tensorflow-estimator=2.7.0=pypi_0
tensorflow-io-gcs-filesystem=0.22.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=py39haa95532_0
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=3.0.0=pypi_0
tk=8.6.11=h2bbff1b_0
tomli=1.2.2=pypi_0
tornado=6.1=py39hb82d6ee_2
traitlets=5.1.1=pyhd8ed1ab_0
typing-extensions=4.0.1=pypi_0
tzdata=2021e=hda174b7_0
urllib3=1.26.7=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=2.0.2=pypi_0
wheel=0.37.0=pyhd3eb1b0_1
wincertstore=0.2=py39haa95532_2
winpty=0.4.3=4
xz=5.2.5=h62dcd97_0
zeromq=4.3.4=h0e60522_1
zipp=3.6.0=pyhd8ed1ab_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.9=h19a0ad4_0
